// src/types/models/activity-schedule-list-response-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Aftermeal API Docs
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface ActivityScheduleListResponseDto
 */
export interface ActivityScheduleListResponseDto {
    /**
     * 
     * @type {number}
     * @memberof ActivityScheduleListResponseDto
     */
    'activityScheduleId'?: number;
    /**
     * 
     * @type {string}
     * @memberof ActivityScheduleListResponseDto
     */
    'name'?: string;
    /**
     * 
     * @type {number}
     * @memberof ActivityScheduleListResponseDto
     */
    'maxParticipants'?: number;
    /**
     * 
     * @type {number}
     * @memberof ActivityScheduleListResponseDto
     */
    'currentParticipants'?: number;
    /**
     * 
     * @type {string}
     * @memberof ActivityScheduleListResponseDto
     */
    'status'?: ActivityScheduleListResponseDtoStatus;
    /**
     * 
     * @type {string}
     * @memberof ActivityScheduleListResponseDto
     */
    'type'?: ActivityScheduleListResponseDtoType;
    /**
     * 
     * @type {string}
     * @memberof ActivityScheduleListResponseDto
     */
    'scheduledDate'?: string;
}

export const ActivityScheduleListResponseDtoStatus = {
    Scheduled: 'SCHEDULED',
    InProgress: 'IN_PROGRESS',
    Canceled: 'CANCELED',
    Completed: 'COMPLETED'
} as const;

export type ActivityScheduleListResponseDtoStatus = typeof ActivityScheduleListResponseDtoStatus[keyof typeof ActivityScheduleListResponseDtoStatus];
export const ActivityScheduleListResponseDtoType = {
    Lunch: 'LUNCH',
    Dinner: 'DINNER'
} as const;

export type ActivityScheduleListResponseDtoType = typeof ActivityScheduleListResponseDtoType[keyof typeof ActivityScheduleListResponseDtoType];